+++ test/onedrive.test.js
@@ -17,10 +17,11 @@
 process.env.HELIX_FETCH_FORCE_HTTP1 = 'true';
 
 const assert = require('assert');
-const nock = require('nock');
+const jose = require('jose');
 const OneDrive = require('../src/OneDrive.js');
 const MockDrive = require('../src/OneDriveMock.js');
 const StatusCodeError = require('../src/StatusCodeError');
+const { Nock } = require('./utils.js');
 
 const AZ_AUTHORITY_HOST_URL = 'https://login.windows.net';
 
@@ -34,8 +35,14 @@ const DEFAULT_AUTH = {
 };
 
 describe('OneDrive Tests', () => {
-  afterEach(() => {
+  let nock;
+  beforeEach(() => {
+    nock = new Nock();
     delete process.env.HELIX_ONEDRIVE_NO_SHARE_LINK_CACHE;
+  });
+
+  afterEach(() => {
+    nock.done();
   });
 
   it('throws when required parameters are not specified.', async () => {
@@ -50,17 +57,21 @@ describe('OneDrive Tests', () => {
   });
 
   it('can encode a share link', () => {
-    assert.deepEqual(OneDrive.encodeSharingUrl('https://onedrive.com/a/b/c/d'),
-      'u!aHR0cHM6Ly9vbmVkcml2ZS5jb20vYS9iL2MvZA=');
+    assert.deepStrictEqual(
+      OneDrive.encodeSharingUrl('https://onedrive.com/a/b/c/d'),
+      'u!aHR0cHM6Ly9vbmVkcml2ZS5jb20vYS9iL2MvZA=',
+    );
   });
 
   it('can encode a share link as url', () => {
-    assert.deepEqual(OneDrive.encodeSharingUrl(new URL('https://onedrive.com/a/b/c/d')),
-      'u!aHR0cHM6Ly9vbmVkcml2ZS5jb20vYS9iL2MvZA=');
+    assert.deepStrictEqual(
+      OneDrive.encodeSharingUrl(new URL('https://onedrive.com/a/b/c/d')),
+      'u!aHR0cHM6Ly9vbmVkcml2ZS5jb20vYS9iL2MvZA=',
+    );
   });
 
   it('can convert a drive item to a uri', () => {
-    assert.deepEqual(OneDrive.driveItemToURL({
+    assert.deepStrictEqual(OneDrive.driveItemToURL({
       id: 'item-id',
       parentReference: {
         driveId: 'drive-id',
@@ -69,7 +80,7 @@ describe('OneDrive Tests', () => {
   });
 
   it('can convert an url to a drive item', () => {
-    assert.deepEqual({
+    assert.deepStrictEqual({
       id: 'item-id',
       parentReference: {
         driveId: 'drive-id',
@@ -78,7 +89,7 @@ describe('OneDrive Tests', () => {
   });
 
   it('can convert an url string to a drive item', () => {
-    assert.deepEqual({
+    assert.deepStrictEqual({
       id: 'item-id',
       parentReference: {
         driveId: 'drive-id',
@@ -87,17 +98,21 @@ describe('OneDrive Tests', () => {
   });
 
   it('returns null for non onedrive urls', () => {
-    assert.equal(OneDrive.driveItemFromURL('https://www.example.com/drives/drive-id/items/item-id'), null);
+    assert.strictEqual(OneDrive.driveItemFromURL('https://www.example.com/drives/drive-id/items/item-id'), null);
   });
 
   it('throws an error for onedrive with wrong format (missing drives)', () => {
-    assert.throws(() => OneDrive.driveItemFromURL('onedrive:/drive-id/items/item-id'),
-      new Error('URI not supported (missing \'drives\' segment): onedrive:/drive-id/items/item-id'));
+    assert.throws(
+      () => OneDrive.driveItemFromURL('onedrive:/drive-id/items/item-id'),
+      new Error('URI not supported (missing \'drives\' segment): onedrive:/drive-id/items/item-id'),
+    );
   });
 
   it('throws an error for onedrive with wrong format (missing items)', () => {
-    assert.throws(() => OneDrive.driveItemFromURL('onedrive:/drives/drive-id/item-id'),
-      new Error('URI not supported (missing \'items\' segment): onedrive:/drives/drive-id/item-id'));
+    assert.throws(
+      () => OneDrive.driveItemFromURL('onedrive:/drives/drive-id/item-id'),
+      new Error('URI not supported (missing \'items\' segment): onedrive:/drives/drive-id/item-id'),
+    );
   });
 
   it('fuzzyGetDriveItem returns folder item when relpath is missing', async () => {
@@ -113,7 +128,7 @@ describe('OneDrive Tests', () => {
       .registerDriveItem('123', '456', data);
     const res = await drive.fuzzyGetDriveItem(folderItem);
 
-    assert.deepEqual(res, [data.value]);
+    assert.deepStrictEqual(res, [data.value]);
   });
 
   it('fuzzyGetDriveItem returns exact item', async () => {
@@ -129,7 +144,7 @@ describe('OneDrive Tests', () => {
       .registerDriveItemChildren('123', '456', data);
     const res = await drive.fuzzyGetDriveItem(folderItem, '/document.docx');
 
-    assert.deepEqual(res, [{
+    assert.deepStrictEqual(res, [{
       extension: 'docx',
       file: {
         mimeType: 'dummy',
@@ -152,7 +167,7 @@ describe('OneDrive Tests', () => {
       .registerDriveItemChildren('123', '456:/publish/en:', data); // this is a bit a hack to trick OneDriveMock
     const res = await drive.fuzzyGetDriveItem(folderItem, '/publish/en/document.docx');
 
-    assert.deepEqual(res, [{
+    assert.deepStrictEqual(res, [{
       extension: 'docx',
       file: {
         mimeType: 'dummy',
@@ -172,7 +187,7 @@ describe('OneDrive Tests', () => {
       .registerDriveItemChildren('123', '456', data);
     const res = await drive.fuzzyGetDriveItem(folderItem, '/document.docx');
 
-    assert.deepEqual(res, []);
+    assert.deepStrictEqual(res, []);
   });
 
   it('fuzzyGetDriveItem returns matching items', async () => {
@@ -197,7 +212,7 @@ describe('OneDrive Tests', () => {
       .registerDriveItemChildren('123', '456', data);
     const res = await drive.fuzzyGetDriveItem(folderItem, '/my-1-document.docx');
 
-    assert.deepEqual(res, [
+    assert.deepStrictEqual(res, [
       {
         extension: 'docx',
         file: { mimeType: 'dummy' },
@@ -247,7 +262,7 @@ describe('OneDrive Tests', () => {
       .registerDriveItemChildren('123', '456', data);
     const res = await drive.fuzzyGetDriveItem(folderItem, '/my-1-document');
 
-    assert.deepEqual(res, [
+    assert.deepStrictEqual(res, [
       {
         extension: 'docx',
         file: { mimeType: 'dummy' },
@@ -295,7 +310,7 @@ describe('OneDrive Tests', () => {
       .registerDriveItemChildren('123', '456', data);
     const res = await drive.fuzzyGetDriveItem(folderItem, '/my-1-document');
 
-    assert.deepEqual(res, [{
+    assert.deepStrictEqual(res, [{
       extension: 'docx',
       file: { mimeType: 'dummy' },
       fuzzyDistance: 5,
@@ -335,16 +350,44 @@ describe('OneDrive Tests', () => {
       clientId: TEST_CLIENT_ID || 'foobar',
       username: TEST_USER || 'test-user',
       password: TEST_PASSWORD || 'test-password',
+      tenant: 'common',
     });
     const me = await od.me();
-    assert.deepEqual(me, expected);
+    assert.deepStrictEqual(me, expected);
   }).timeout(5000);
 
-  it('uses global share link cache by default', async () => {
-    const scope0 = nock(AZ_AUTHORITY_HOST_URL)
+  it('can authenticate against a resource', async () => {
+    nock(AZ_AUTHORITY_HOST_URL)
       .post('/common/oauth2/token?api-version=1.0')
-      .reply(200, DEFAULT_AUTH);
-    const scope1 = nock('https://graph.microsoft.com/v1.0')
+      .reply(200, {
+        token_type: 'Bearer',
+        refresh_token: 'dummy',
+        access_token: 'dummy',
+        expires_in: 81000,
+      });
+
+    const od = new OneDrive({
+      clientId: 'test-client-id',
+      clientSecret: 'test-client-secret',
+      resource: 'test-resource',
+      tenant: 'common',
+    });
+    const resp = await od.getAccessToken();
+    delete resp.expiresOn;
+    assert.deepStrictEqual(resp, {
+      _authority: 'https://login.windows.net/common',
+      _clientId: 'test-client-id',
+      accessToken: 'dummy',
+      expiresIn: 81000,
+      refreshToken: 'dummy',
+      resource: 'test-resource',
+      tokenType: 'Bearer',
+    });
+  });
+
+  it('uses global share link cache by default', async () => {
+    nock.loginWindowsNet();
+    nock('https://graph.microsoft.com/v1.0')
       .get('/shares/u!aHR0cHM6Ly9vbmVkcml2ZS5jb20vYS9iL2MvZDA/driveItem')
       .reply(200, {
         id: 'some-id',
@@ -354,6 +397,7 @@ describe('OneDrive Tests', () => {
       clientId: 'foobar',
       refreshToken: 'dummy',
       localAuthCache: true,
+      tenant: 'common',
     });
     const item1 = await od.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d0');
     assert.deepStrictEqual(item1, {
@@ -361,17 +405,13 @@ describe('OneDrive Tests', () => {
     });
     const item2 = await od.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d0');
     assert.strictEqual(item1, item2);
-
-    scope0.done();
-    scope1.done();
   });
 
   it('share link cache can be disabled via option', async () => {
-    const scope0 = nock(AZ_AUTHORITY_HOST_URL)
+    nock(AZ_AUTHORITY_HOST_URL)
       .post('/common/oauth2/token?api-version=1.0')
-      .twice()
       .reply(200, DEFAULT_AUTH);
-    const scope1 = nock('https://graph.microsoft.com/v1.0')
+    nock('https://graph.microsoft.com/v1.0')
       .get('/shares/u!aHR0cHM6Ly9vbmVkcml2ZS5jb20vYS9iL2MvZDAx/driveItem')
       .twice()
       .reply(200, {
@@ -383,26 +423,23 @@ describe('OneDrive Tests', () => {
       refreshToken: 'dummy',
       localAuthCache: true,
       noShareLinkCache: true,
+      tenant: 'common',
     });
     const item1 = await od.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d01');
     assert.deepStrictEqual(item1, {
       id: 'some-id',
     });
     const item2 = await od.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d01');
-    assert.deepEqual(item1, item2);
+    assert.deepStrictEqual(item1, item2);
     assert.notStrictEqual(item1, item2);
-
-    scope0.done();
-    scope1.done();
   });
 
   it('share link cache can be disabled via env', async () => {
     process.env.HELIX_ONEDRIVE_NO_SHARE_LINK_CACHE = true;
-    const scope0 = nock(AZ_AUTHORITY_HOST_URL)
+    nock(AZ_AUTHORITY_HOST_URL)
       .post('/common/oauth2/token?api-version=1.0')
-      .twice()
       .reply(200, DEFAULT_AUTH);
-    const scope1 = nock('https://graph.microsoft.com/v1.0')
+    nock('https://graph.microsoft.com/v1.0')
       .get('/shares/u!aHR0cHM6Ly9vbmVkcml2ZS5jb20vYS9iL2MvZDE/driveItem')
       .twice()
       .reply(200, {
@@ -413,24 +450,20 @@ describe('OneDrive Tests', () => {
       clientId: 'foobar',
       refreshToken: 'dummy',
       localAuthCache: true,
+      tenant: 'common',
     });
     const item1 = await od.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d1');
     assert.deepStrictEqual(item1, {
       id: 'some-id',
     });
     const item2 = await od.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d1');
-    assert.deepEqual(item1, item2);
+    assert.deepStrictEqual(item1, item2);
     assert.notStrictEqual(item1, item2);
-
-    scope0.done();
-    scope1.done();
   });
 
   it('share link cache can be supplied via opts', async () => {
-    const scope0 = nock(AZ_AUTHORITY_HOST_URL)
-      .post('/common/oauth2/token?api-version=1.0')
-      .reply(200, DEFAULT_AUTH);
-    const scope1 = nock('https://graph.microsoft.com/v1.0')
+    nock.loginWindowsNet();
+    nock('https://graph.microsoft.com/v1.0')
       .get('/shares/u!aHR0cHM6Ly9vbmVkcml2ZS5jb20vYS9iL2MvZDI/driveItem')
       .reply(200, {
         id: 'some-id',
@@ -443,6 +476,7 @@ describe('OneDrive Tests', () => {
       refreshToken: 'dummy',
       localAuthCache: true,
       shareLinkCache,
+      tenant: 'common',
     });
     const item1 = await od.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d2');
     assert.deepStrictEqual(item1, {
@@ -456,8 +490,165 @@ describe('OneDrive Tests', () => {
         id: 'some-id',
       },
     });
-    scope0.done();
-    scope1.done();
+  });
+
+  it('resolves the tenant from a share link and caches it', async () => {
+    nock(AZ_AUTHORITY_HOST_URL)
+      .get('/onedrive.onmicrosoft.com/.well-known/openid-configuration')
+      .reply(200, {
+        issuer: 'https://sts.windows.net/c0452eed-9384-4001-b1b1-71b3d5cf28ad/',
+      })
+      .post('/c0452eed-9384-4001-b1b1-71b3d5cf28ad/oauth2/token?api-version=1.0')
+      .twice()
+      .reply(200, {
+        token_type: 'Bearer',
+        refresh_token: 'dummy',
+        access_token: 'dummy',
+        expires_in: 81000,
+      });
+    nock('https://graph.microsoft.com/v1.0')
+      .get('/shares/u!aHR0cHM6Ly9vbmVkcml2ZS5jb20vYS9iL2MvZDI/driveItem')
+      .twice()
+      .reply(200, {
+        id: 'some-id',
+      });
+
+    const tenantCache = new Map();
+    const od1 = new OneDrive({
+      clientId: 'foobar',
+      refreshToken: 'dummy',
+      localAuthCache: true,
+      tenantCache,
+      noShareLinkCache: true,
+    });
+    const item1 = await od1.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d2');
+    assert.deepStrictEqual(item1, {
+      id: 'some-id',
+    });
+
+    const od2 = new OneDrive({
+      clientId: 'foobar',
+      refreshToken: 'dummy',
+      localAuthCache: true,
+      tenantCache,
+      noShareLinkCache: true,
+    });
+    const item2 = await od2.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d2');
+    assert.deepStrictEqual(item1, item2);
+
+    assert.deepStrictEqual(Object.fromEntries(tenantCache.entries()), {
+      onedrive: 'c0452eed-9384-4001-b1b1-71b3d5cf28ad',
+    });
+  });
+
+  it('resolves the tenant from a sharepoint share link and caches it', async () => {
+    nock(AZ_AUTHORITY_HOST_URL)
+      .get('/adobe.onmicrosoft.com/.well-known/openid-configuration')
+      .reply(200, {
+        issuer: 'https://sts.windows.net/c0452eed-9384-4001-b1b1-71b3d5cf28ad/',
+      })
+      .post('/c0452eed-9384-4001-b1b1-71b3d5cf28ad/oauth2/token?api-version=1.0')
+      .reply(200, {
+        token_type: 'Bearer',
+        refresh_token: 'dummy',
+        access_token: 'dummy',
+        expires_in: 81000,
+      });
+    nock('https://graph.microsoft.com/v1.0')
+      .get('/shares/u!aHR0cHM6Ly9hZG9iZS1teS5zaGFyZXBvaW50LmNvbS9hL2IvYy9kMg=/driveItem')
+      .reply(200, {
+        id: 'some-id',
+      });
+
+    const tenantCache = new Map();
+    const od1 = new OneDrive({
+      clientId: 'foobar',
+      refreshToken: 'dummy',
+      localAuthCache: true,
+      tenantCache,
+      noShareLinkCache: true,
+    });
+    const item1 = await od1.getDriveItemFromShareLink(new URL('https://adobe-my.sharepoint.com/a/b/c/d2'));
+    assert.deepStrictEqual(item1, {
+      id: 'some-id',
+    });
+    assert.deepStrictEqual(Object.fromEntries(tenantCache.entries()), {
+      adobe: 'c0452eed-9384-4001-b1b1-71b3d5cf28ad',
+    });
+  });
+
+  it('resolves the tenant from a share link and ignores cache', async () => {
+    nock(AZ_AUTHORITY_HOST_URL)
+      .get('/onedrive.onmicrosoft.com/.well-known/openid-configuration')
+      .twice()
+      .reply(200, {
+        issuer: 'https://sts.windows.net/c0452eed-9384-4001-b1b1-71b3d5cf28ad/',
+      })
+      .post('/c0452eed-9384-4001-b1b1-71b3d5cf28ad/oauth2/token?api-version=1.0')
+      .twice()
+      .reply(200, {
+        token_type: 'Bearer',
+        refresh_token: 'dummy',
+        access_token: 'dummy',
+        expires_in: 81000,
+      });
+    nock('https://graph.microsoft.com/v1.0')
+      .get('/shares/u!aHR0cHM6Ly9vbmVkcml2ZS5jb20vYS9iL2MvZDI/driveItem')
+      .twice()
+      .reply(200, {
+        id: 'some-id',
+      });
+
+    const od1 = new OneDrive({
+      clientId: 'foobar',
+      refreshToken: 'dummy',
+      localAuthCache: true,
+      noTenantCache: true,
+      noShareLinkCache: true,
+    });
+    const item1 = await od1.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d2');
+    assert.deepStrictEqual(item1, {
+      id: 'some-id',
+    });
+    const od2 = new OneDrive({
+      clientId: 'foobar',
+      refreshToken: 'dummy',
+      localAuthCache: true,
+      noTenantCache: true,
+      noShareLinkCache: true,
+    });
+    const item2 = await od2.getDriveItemFromShareLink('https://onedrive.com/a/b/c/d2');
+    assert.deepStrictEqual(item1, item2);
+  });
+
+  it('sets the access token an extract the tenant', async () => {
+    const keyPair = await jose.generateKeyPair('RS256');
+    const bearerToken = await new jose.SignJWT({
+      email: 'bob',
+      name: 'Bob',
+      userId: '112233',
+      tid: 'test-tenantid',
+    })
+      .setProtectedHeader({ alg: 'RS256' })
+      .setIssuedAt()
+      .setIssuer('urn:example:issuer')
+      .setAudience('dummy-clientid')
+      .setExpirationTime('2h')
+      .sign(keyPair.privateKey);
+
+    const od = new OneDrive({
+      clientId: 'foobar',
+      refreshToken: 'dummy',
+      localAuthCache: true,
+      noTenantCache: true,
+      noShareLinkCache: true,
+    });
+    od.setAccessToken(bearerToken);
+
+    const accessToken = await od.getAccessToken();
+    assert.strictEqual(accessToken.accessToken, bearerToken);
+    assert.strictEqual(accessToken.tenantId, 'test-tenantid');
+    assert.strictEqual(od.tenant, 'test-tenantid');
   });
 
   it('propagates errors', async () => {
@@ -481,6 +672,7 @@ describe('OneDrive Tests', () => {
       clientId: TEST_CLIENT_ID || 'foobar',
       username: TEST_USER || 'test-user',
       password: TEST_PASSWORD || 'test-password',
+      tenant: 'common',
     });
     await assert.rejects(od.me(), new StatusCodeError('wrong input', 400));
   }).timeout(5000);